Generate fresh names when parsing processes
Term.fresh_name could not be implemented there, I moved it to Theory
where the global symbols table is maintained. In the long run we should
rethink the organization of Term and Theory, which has drifted over
time, and move global macros to the same table of declarations as other
symbols.

Fix #26, ref #2.