Parse action symbols
Users can (and must) write actions as a(i,j) where a is an action
symbol. This is what we manipulate in Theory terms. Symbols are turned
to structured actions when converting to Term terms, which have not
changed. Actions are still printed in their structured form, which
should disappear in the next step.

Tests have been updated following this change. Additionnally I have
removed two useless hypotheses in one euf test.