Fix Term.macro unfolding
Term macros (which correspond to let definitions) only make sense for
actions of a given length. This information is now part of the macro
definition, and used properly by Euf.

This fixes #31 but made me realize that the current process preparation
code does not maintain a list of input variables that corresponds to
the final actions that are produced. I had to enrich the process
preparation, but this code is probably still not perfect.