Print action symbols + introduce Symbols
This commit completes the introduction of action symbols, which are
now used when printing actions. This required the introduction of
a new architecture for global definitions, which will eventually all
use the Symbols module: it will provide a centralized place for all
global declarations, making it possible to have specific kinds of
symbols with definitions depending on Action, Term, etc.