Fix #38
The issue was not the null system, but the presence of message
variables. The syntactic side condition did not fail on them because
it used to apply only to actions and not to the direct terms of
the fact on which euf is applied, which has been fixed in c5e4ef95.

I updated examples/euf_null.mbc to clarify the problem, and moved
it to tests/fail for regression testing.