More scalable tactic syntax trees
Introduce a notion of tactic AST with simple named tactics at
their leaves. This is what the parser produces, and what is
pretty-printed: these parts of the code are now independent of
the list of actually declared tactics. To make this possible
I introduced a generic notion of tactic argument.

Prover.Prover_tactics is now the single place where a new
tactic should be declared. In the future, this declaration will
come with documentation for the user.

The tactics ASTs are actually independent of the judgment type:
we have generics ASTs in Tactics, and they are instantiated in
Prover.AST for Judgment.t tactics.

The type of tactics has also been simplified with a purely
abstract return type. This is a stronger type discipline than
the former polymorphic type, which also could not be maintained
when storing tactics in the Prover_tactics hash table.

All these changes should have no effect on the user interface.
There is a minor exception that I have no investigated in detail,
in "tests/ok/imply.mbc".