Simplify and generalize goal simplification
Add the introduction of true and empty existentials in the implicit
simplification tactic, which makes the simplify and remove_finished
steps superfluous.

Fix the problem on tests/ok/imply.mbc introduced in aa8ef73e:
in that commit I had removed a goal simplification performed at
the end of the intro tactic (it was rather ad-hoc, and created
a circular dependency between Prover and Trace_tactics); in the
test, an existential remained, was simplified, but because the
simplification was not iterated the goal could not be simplified
enough to be removed.