Fix exists & update tests and examples using it
The tactic had been renamed, and this change had not been propagated.

Moreover, the tactic did not check the length of its substitution,
and it applied (incorrectly) on non-empty existentials when invoked
(though intro or anyintro) with an empty substitution. This broke
examples/lak_tags.mbc, which is now fixed.