Add happens(_) predicate
The new predicate is handled in tactics through the extension of
conjunction/disjunction_to_atoms, which now extracts happens atoms.
The addition of these atoms to (the left-hand side of) judgments
has the effect of adding the condition of the action to the judgment.
Outputs and states are not added at this point, but are (or will be)
treated through the macro system.

Update LAK and MW examples, as well as tests. Test euf_42 has been
trivialized because I could not make sense of it: I did not remove
it because an upcoming change to euf will be tested on it.