Remove abusive addition of action condition in euf
Fix #54. This has not broken any example. There might be cases
where it would make a difference: in these cases we will have to
rely on an improved constraint solver that will be able to derive
happens(B) from happens(A) and B<=A (where A and B are action names).