Fix euf
Rather than parsing this tactic specifically, allow Theory.terms as
tactic arguments. A hypothesis name is going to be parsed as a variable,
which euf then interprets as a hypothesis name. This is not very robust
as the Theory.term parser might interpret some hypothesis names
as specific terms. We'll have to discuss various possibilities for
better solutions.