Pretty printing and parsing of formulas, etc.
Do not reorder variables listed in quantifiers, fix #60.

Instead, group consecutive variables when they have the same type, for
a more compact display of quantifiers.

Allow short forms for quantifiers: the comma is optional, the
parentheses also when there is a single variable.

Print types when displaying the environment of a sequent.

Introduce unit tests, notably checking that pretty-printing and parsing
are inverses of each other. Fix printing of propositional constants, and
simplify/uniformize box treatment for various connectives.