Timestamp case and left existential tactics
Add "case" and "existsleft" tactics, with corresponding tests.
As it is, the "case" tactic is unsound, because it does not cover epsilon.

Generalize "splitleft" to work on arbitrary disjunctions and not only
disjunctions of atoms.

Improve "examples/wip/collisions.mbc" thanks to these tactics, though it
is not yet satisfactory: it shows limitations in the treatment of
macros, and a case where eqnames does not seem to do its job.

Also did some slight API change in Sequent, and some re-indenting.