[WIP] Trying to remove bformula
Sequent now stores list of trace_atoms rather than trace_formulas.
We lost a bit of automation, as a disjunction in the conclusion is not added
as a conjunction of negations to the set of hypothesis when trying to derive
false. Note that this is similar to message hypothesis.
I have difficulties in removing fact, as I do not know exactly the
dependencies in process.ml.