Implement term as gadt. #64
The sort of a variable or a term is now explicit. We use ESort : 'a sort -> esort
 and EVars of 'a var to wrap explicit types. There is a type 'a term, which can be
instantied for Term.messages (old Term.term) or Term.timestamp.
There is no longer multiple constructor inside messages or timestamps form variables.
There is a single constructor for substitutions, ESubst : 'a term * 'a term -> esubst.

<! Still some clean up to do. !>
	- Remove open of Vars,Sorts and Terms
	- Remove dead code
	- Comments