New tactic error management and printing interface. #65
Tactics exceptions are now only transformed to string inside the main loop.
A test mode allows to keep the raised expection, and thus, perform unit
testing on tactics. Cf the test at the end of main.ml.

The printing interface has been revamped, so that the output channel can depend
on the running mode. It unifies the management of the output API, for instance
for Proof General.

Import changes:
    - main.ml is now parserbuf.ml
    - metabc.ml is now main.ml
    - metabc.ml only contains the call to the Main.main ()
        else, adding tests inside main.ml runs the interactive tool
    - All printing should be done using Printer.pr or Printer.prt

TODO: all failing tests should be done using alcotest, matching on the
expected exception.