New term operators, updating Theory, etc.
Add if-then-else and try-find to terms, as well as operators for biterms:
diff(_,_), left(_) and right(_). The last two allow to override projections,
which will be useful for equivalences: equivalences are given as bi-frames,
representing the equivalence of the left and right projections; sometimes one
wants a left projection on the right (or the converse) which will be allowed
thanks to the left(_) and right(_) operators.

Propagate these changes in a few places. Start updating Theory to merge
formulas and terms.