Adapt Theory to mirror Term
Theory terms and formulas are merged, and we introduce biterm operators
as well as if-then-else and try-find operators.

As for the modification in Term, this actually simplifies the code by
merging functionalities for term and formulas. In this case, it also
lead to an important cleanup of the typechecking and conversion functions:
typechecking is now a byproduct of conversions, and conversions to various
type of terms are now unified thanks to GADTS.