Support for equivalence goals in prover input
It is now possible to declare equivalence goals. The syntax is:

equiv <name> <variables> : <bi-terms>.

Unlike for trace goals, the goal name is mandatory to avoid parsing conflicts.
Bi-terms can be messages or formulas. The variables are optional, written
as in quantifiers, and should contain all free variables.

A first test gives a few examples.

This is not very user-friendly, but it allows to get started.

Also extend parser: the else branches of if-then-else and try-find are now
optional in terms (when omitted they stand for "... else zero").