Extend FA, restrict ITE and Find to messages
These operators are meant only for messages in practice, and restricting
them in the type definition makes pattern matching more precise, easier
to program with.

Add support for ITE in the "fa" tactic, and diff operator when parsing
formulas. The tactic is still rudimentary: the existing test shows that
it produces biterms with redundant diff operators.