First tactic using reachability in equivalence. Expand tactic in equiv.
Implementation of the noif tactic, which applied to a ite(b,c,d), creates the
goals b^L => False, b^R => False, and same diffe equivalence using d.
For the moment we can only create goals where a single action contain diff.
A None was added to the projection type, to simplify the usage of sequents.