Improve conversion errors. Cf #83.
When trying to convert for any types, we should only catch type_error exceptions.
I am not sure how we should improve the structure of the conversion, notably for
macros it can be hard to say if a symbol of ill-typed or undefined, because we
have a very ad-hoc way to handle them.