Multiple system definitions. #75
One can now define multiple systems, where bi-actions are registered under a system name.
Base_system allows to model either a process or a bi-process given a set of actions and
a projection. Given two base_system, an Action.system allows to model either a process or a
bi-process, built from two possible disinct set of actions (but with the same shapes).