Fix EUF wrt. bound variables & key indices constr.
Continue fixing the issue with bound variables, this time in direct
cases, ref #81. I realized that constraints on key indices were missing
even without bound variables: I fixed this as well. This has required
some slight changes in the structure of the code, and I now wonder
whether the treatment of direct and indirect cases could be unified.