Introducing DDH tactic
In any goal, if the system and the frame only use correcly the name symbol
a, b, and c, one can add to the assumptions of the system that g^a^b \sim g^c
and g^b^a \sim g^c. This is the simplest form of DDH possible.