Definition of tag formula for hash integrated to hash definition. Close #89
The syntax `hash h with oracle formula` allows to declare a tag formula.
The tag formula are not stored inside Symbols data, as it would cause a
circular dependency. They are stored inside the proved goal, but with a name
definied uniquely from the hash function name, and prefixed with a %, so that
it cannot clash with a user defined axiom, or be applied in a proof.