Cleanup exception uses in trace tactics
In order to be usable and debuggable with combinators, tactics must
follow two guidelines:

 - Only hard failures can be raised, and only when arguments are
   improper regardless of the goal.

 - Calls to continuations must be tail calls.

The programming style used to enforce this calls for a new API for
registering tactics.

I have not been completely exhaustive: at least "euf" and "apply" are
still incorrect, but it matters less because they are not used in
the "simpl" combined tactic.