Introduce newsimpl
Introduce newsimpl which works as a replacement of simpl: it avoids all
forms of backtracking and attempts to be smarter when dealing with
splits. Performances are as before, except for extreme cases à la
nobranch.mbc: for now the old simpl remains present in the code
for further analysis.

The new tactic behaves slightly differently than the old one: instead of
splitting everything and trying to leave at most one subgoal opened
(with backtracking) it splits conjunctions one by one, controlling the
remaining subgoals on both sides. For instance, it reduces True & (True
& (P & Q)) into (P & Q) while the old tactic would leave the whole
conjunction untouched: it is unclear to me what is most desirable.

Slightly enrich intro to simplify the definition of intros, anyintro and
(old) simpl. Also slightly cleanup simpl.

Introduce andthen with cut, which removes backtracking.

Fix the meaning of try_tac and test it.