First attempt at transitivity tactic inside reachability prover.
Over some systems, if it holds that forall i:index, cond@A(i) <=> f(i), we can
in effect replace in any proof the system by the one where cond@A(i) has been
syntactically replaced by f(i). The susbtitutesysten tactic allows such a thing.
Actions.ml now provide an interface to create a clone of a system where such a
substitution was applied.
The parsing of the tactic and the syntax is uber ugly, feedback and ideas welcome.