Finalize a4e0cb716e9aae67ce04c38e3e95a5b151adc690
Add regression test. Also improve a bit the tactic to avoid adding
trivial and redundant equalities. Document get_all_terms and collision
to clarify that they only consider toplevel terms: the tactic could
be improved to consider subterms, but it might not be worth it since
it will disappear in the long run.