Improve fadup
The fadup verification did not respect scoping. To fix this, I
introduced Iter.fold which is analogous to Iter.iter but allows
to pass a "state" value as is usual fold functions.

Also generalize fadup to decouple the treatment of conjunctions
and implications from quantifiers: the theory allows it has to be
updated.

Finally, allow fadup do work on sequences of formulas, and also
allow binding construct inside terms and formulas in fadup: the
existing mechanisms from Iter nicely avoid variable capture
issues.

It may be possible to make the tactic more flexible regarding the
topevel conjunction with an exec atom, if we could initialize the
verification with an empty list of known executable timestamps.

Update wip/mw-equiv which benefits from this generalization.