Add new case tactic, complete mw-seq
The trace tactic "case" now allows to perform case analysis on a
conditional term (if-then-else or try-find) or a macro that expands
to a conditional term.

The trace tactic "fa" seems to be subsumed by this new capability,
though it is more convenient as it does not require to write a
potentially large term when invoking the tactic.

This is used to complete the missing cases in examples/wip/mw-seq.mbc.