Generic system for tactic declarations. #104
Unifies the tactic declaration, by providing new register methods that take cares
of tail-recursion and types conversions:
 - creates an TacticArgs modules, with explicit types for arguments
 - instantiate multiple register functions, that simply expects functions over
   judgment and a type specification. See the trace tactic left_intros for a simple
   use case.

The down-side is that we have multiple registration function depending on the
number of arguments. The up-side is that in the trace tactics, everything is
cleaned up. Some tactics, due to the complexity of their arguments, are still
using register_general.

If the system is convincing, it remains to port most of the equivalence tactics.