Add global option declaration in prover, and add oracle for hash
Global options can be now defined and stored in the prover states.
It has been used to save oracle formulas associated to hash or
signature.

Could be used for other things later one.

Oracle for hash supports all previous tactics, where collision
resistance does not care about the tagged formula, euf does
as with the signatures, and prf adds to the condition the negation
of the tag.