Improve parsing of optional args for tactics
It is now possible to have multiple optional arguments.

In order to preserve some tests using ifcond, I had to introduce some
backtracking when trying to parse an optional argument. This was not
present in the original code thanks to the assumption that there is
at most one optional argument.