Fix SSC for symmetric encryptions assumptions.
In symmetric encryptions assumptions (ENCKP, INDCCA and INTCTXT),
the attacker is given an encryption oracle. It means that we must
check that all encryption with the key can be produced by a call
to the encryption oracle. In particular, it means that a name
used as randomness for an encryption must only occur in such position,
and two cyphertexts with the same random must have equal plaintext.

As an overapproximation, we implemented this as a SSC, where we
ask that:
 - the randoms used in encryption appear nowhere else;
 - a random is never used in two distinct encryption;
 - when a random is used inside a replicated action, the random depends
   on more indices than the plaintext does.