Register empty dummy action
A let-definition before any input used to be registered with a dummy
input variable generated by Process.prepare. Such definitions are now
registered with an empty list of inputs. The difference is irrelevant
since no input variable can actually be reference from the definition's
body, but the newly used empty dummy action was not properly registered
in shape_to_symb and could thus not be converted to a term.