Handle action reuse in process parsing
When an already existing action symbol is forced by Action.register,
it should be used in the rest of the translation. This change replicates
the old behaviour, but it seems incomplete: the old action name may
still be used in the description that has just been registered.
This may result in weird behaviours, and impossible proofs.