Revamp fail test facilites -> CheckFail tactical and Abort #119
Introduces the tactical checkfail, which allows to check if a tactic
fails with the expected exception.
Abort allows to abort a proof (only if it is not finished).

Using those facilities, all the tests pertaining to euf where
centralized in tests/fail/euf.mbc.

Ultimately, most alcotest failures should be handled this way, especially
for checking SSC of inference rules.

It remains to think about how this system could be interfaced with the
manual.