fixed several memoisation functions
on my machine, this caused a performance improvement of around 12%

the memoisation was done through ephemerons built over tuples,
where the tuples appeared nowhere in the prover: this caused entries to
be cleared prematurely.

Now, we use Ephemeron.K2 or Ephemeron.Kn where possible, and
Hashtbl.Make otherwise. In the latter case, the entries are never
cleared from the memoisation table, which may cause some memory leaks.