Global goals and axioms, more uniform APIs
Equivalences in global formulas must now be written as equiv(biframe)
or ~(biframe). The old "equiv" goals are still available, allowing
to write just the biframe. For more complex cases, "global goal" must
be used. It is now also possible to postulate a "global axiom".

The syntax (parsing) of global formulas is otherwise unchanged, but
should be adapted next to differentiate its quantifiers from local
meta-formula quantifiers. Regarding pretty-printing, quantifiers
are uppercase but equivalences are still shown as bare biframes.

Examples and tests are updated to the new syntax. This made it possible
to uncomment several planned tests, but one of them breaks.

As part of this change, attempt to rationalize our APIs and handle
more uniformly our two kinds of meta-formulas:
 - More uniform API for the creation of local and global goals.
 - Remove Decl.p_goal_reach, Goal.lemma_g and Goal.p_goal_form
   which were both hard to use, slightly redundant and had misleading
   names.
 - The new API uses the following terminology:
   * goal is for a internal proof goal, ie. a sequent;
   * conclusion is for the right hand-side formula of a sequent;
   * hypotheses is for the left hand-side formulas of a sequent;
   * statement is for a formula with enough context to understand it
     (eg. system, bound variables) cf. Goal.statement;
   * assumption is for a statement that is assumed to hold,
     ie. an axiom or an hypothesis in a goal.
   For instance, the *_lemma functions in Goal are renamed to *_statement.
   In Prover and Sequent, they are renamed to *_assumption.
 - Rename Theory.equiv_form type into global_formula, add locations to it.
   More renaming of the sort remains to be done, also for local formulas
   which are sometimes called reachability, sometimes trace formulas.