Fix and improve reach_equiv
The standard system compatibility check performed by get_assumption is
inadequate for reach_equiv. Add the possibility to disable this check,
and implement a custom test in the reach_equiv tactic. Also strengthen
the default check when assumption is global.

The tactic now properly change the local sequent's system after
rewriting and allows rewriting in both directions: this is dictated by
the system expression; it will be more flexible (and perhaps more
predictable) in the future to let the user specify the direction.