Fix and improve reach_equiv, update LFMTP example
Fix and extend support for input macros in reach_equiv. Inputs were
changed into frames by mistake and I had not noticed it in the tests.
Moreover, the special handling of inputs in only justified for
timestamps that happen. Updated the tests and adapt them to a style
which checks that inputs do not become frames unexpectedly.

Add support timestamp macros in reach_equiv.

LFMTP example now only has admits waiting for the completion of the
improvement of PRF using refined macro support computations.