Bullet support
Currently +, * and any infix operator can be used as a bullet: we
probably want to be more restrictive in the future.

Differentiate parser function for ProofMode and other modes to avoid
conflicts between bulletted tactics and declarations.

All of the bullet infrastructure is in the bullet module. That module
also supports braces but this is currently unused.

Add bullets in BasicHash example.