canauth.sp requires alt-ergo.2.4.1
Alt-ergo is only installed with 2.2.0 after the pin of alt-ergo-lib to dev,
which is not enough for canauth.sp to work. I am trying to install alt-ergo
while ignoring this downgrade.

We are having a crazy dependency loop, that hopefully should be cleaned up
as soone as alt-ergo-lib is up to date in opam....