Support for braces, etc.
Add support for braces in proof scripts, to enclose the part of a
proof-script that corresponds to a given subgoal.

Update BasicHash and Feldhofer examples as a demonstration, also indenting
comments as a demonstration of what I would like to see elsewhere. The HTML
output is now broken on these example: there is a mismatch between
commands and their outputs.

Braces are parsed, just like bullets, together with the next tactic.
This is awkward (as with bullets) and limiting (because we cannot close
a brace before QED or ABORT without also enriching their parsing).
For the record I tried a more elegant strategy, separating bullets,
braces, and tactics: this allows to merge back top_proofmode and
interactive in parser.mly without conflicts, and works in batch mode
but not in interactive mode where the mismatch between Squirrel steps
(two steps for a bullet followed by a tactic) and ProofGeneral steps
(one step only) messes things up.

Also add support for "-" bullet and check for missing bullets before
evaluating tactics.