Generalize treatment of bullets, fix undo bug
It is now possible to input "}." anywhere in a proof, including at the
very end. Any combination of bullets and braces can prefix a tactic,
or can be executed on its own, though many combinations are actually not
going to be legitimate or useful.

Also decouple eval_tactic and is_proof_completed.