Generalize systems and expressions
This commit brings significant changes. It compiles but many things are
broken, though examples/basic-hash.sp passes.

Systems are now general multi-systems. System expressions can refer
to sets of systems, including the set of all systems and also sets of
compatible systems. Subtyping is used to statically identify system
expressions for which particular operations are available.

Sequents (and several other notions and functions) carry a
SystemExpr.context indicating with respect to which system(s)
each kind of atom (local formula or equivalence) should be interpreted.

The semantics of diff terms changes a little bit: subterms are now
interpreted in the projected systems and cannot contain diff operators
themselves anymore.

The code has been adapted in most places to support this generalization,
but some bugs need to be fixed. The next step is to restore something
close to the old system, with the existing soundness problems, at which
point merging can be considered.