Expand global macros all the time
Expansion is allowed for single systems where the global macro is not
actually defined, to allow the formation of multiple systems where
global macros are only used in some projections, which we use in several
examples and probably want to keep using for readability. In such cases
the macro expands to empty.

This restores the old behavior, but avoids the old expansion using the
default body that resulted in ill-defined terms.