More convenient system annotation for gprf lemma
The former annotation was more general but did not pass the
compatibility check when used in ikeV1_psk. We will be able to
generalize the annotation after some more work on use and apply,
which should eventually integrate the compatibility check.

Also remove SystemExpr.update which was intended for encapsulating
the effect of nested annotations on contexts, which is useless for now.
Its current uses have been replaced by dedicated functions with clearer
specification.