Cleanup rewrite equiv
More meaningful verification of system annotations, made possible by
passing a full SE.context rather than a SE.expr to the tactic.

In order to cover existing examples, allow the tactic to work on
reachability sequents whose set annotation is not a singleton, but
a list where all elements correspond to the same single system.
Added a test to cover this case.

As part of this cleanup, a new source of unsoundness has been found,
which will need to be fixed (see TODO in the code).