Work with global hypotheses more often
When converting an equivalence sequent into a trace sequent, keep all
global hypotheses untouched, instead of converting local formulas from
global to local hypotheses at this time. This was not required by any
example but simplifies the code and could be useful in the future.

This change forced several slight generalizations to preserve working
examples:
- Generalize "assumption" tactic to take global hypotheses into account.
- Similarly generalize LowTraceSequent.get_atoms_of_fhyps.
- Generalize "project" to preserve local formulas as global hypotheses.
- Generalize Sequent.get_assumption to silently convert local formulas
  from global hypotheses to local ones, which makes it possible e.g. to
  rewrite using a global hypothesis.
- Add "localize" tactic to local formulas from global to local hypotheses.

Add tests for most of these new behaviours.