Add "trans" and "sym" tactics
Transitivity is not as general as one might expect. However this
restricted form is useful in practice and is probably the best we
can do without having to pass explicit terms to the tactic (which
would lead to proofs that are difficult to maintain). More behaviours
can be added later on anyway.