allow to unfold operator more often
It used to be the case that operators' type arguments had to be
inferred directly from the the operator (term) arguments.

This is no longer the case, because type arguments are inferred after
typing the parsed terms, and are then attached to the term.

Operator unfolding could only be done if type inferrence
succeeded. This is always the case now.