Remove help system for tactics
A "help" command remains which points to the online user documentation.

By the way, rework the ProverLib types for prover inputs (and the parser)
to clarify when "undo" is accepted or not. Also remove "open Driver" in
Main, which shows how untangled the two modules are, and allowed to
identify an incorrect local redefinition of stat_filename.