Support for benchmarking alternatives to Constr
Use more TraceSequent.query, which groups model computation and
querying, and allow benchmarking of alternatives to Constr at this
level. Technically, the benchmarking takes places on a generalization
of query which also encompasses TraceSequent.constraints_valid.

Benchmarking can be enabled for Smt by setting BENCHMARK_SMT in
the environment.

Benchmark data is logged in BENCHMARK_DIR if it is set.

This mechanism could be extended to more uses of Constr, but it needs
to be used only when the model computation and model analysis are
grouped.

TODO:
- Position information in benchmarking logs is currently very messy.