Documentation + cleanup
Remove mli for LowSequent module which is signature-only.

Remove some references to avoiding circular dependencies involving Prover.
They are not relevant anymore, but it is not obvious to get rid of
current stratification with Low{Trace/Equiv}Sequent and {Trace/Equiv}Sequent
as the latter implement Sequent which features a cast function
to_general_sequent returning Low{Trace/Equiv}Sequents.