Extend benchmarking facility to auto and autosimpl
Only TraceTactics.auto{_simpl} are concerned for now.
SMT-based alternatives are registered if BENCHMARK_SMT is set,
and benchmarking data is logged if BENCHMARK_DIR is set; different
files are produced for the different benchmarks. It is currently
not possible to selectively run a single benchmark.

Also add goal name and bullets to position information in benchmark
data.