Fix lexical feedback (re)initialization bug
The lexical feedback flag for parsing keywords as identifiers
was not properly reset upon error.

Thanks to Stanislas for the minimal bug report. For the record:
```
system null.

lemma _ : true.
Proof.
  smt ~pure=true.
Qed.
```
A parse error is expected for the smt tactic. After that, replacing
the line with just `smt.` should work, but did not because `smt` was
understood as an identifier (and the tactic was improperly invoked).