Fix more CPS errors and fix tests
TD;LR Don't mess with CPS, double-check future changes!

The previous fix to wrap_fail uncovered errors in tests, fixed in this
commit. In short, the CPS violation in wrap_fail broke the tactic
combinators and caused some tests to pass for bad reasons. The tests
are now fixed and made more precise by the way in some cases.

Also fix, by the way, another CPS violation in re_raise_tac. This did
not seem to have consequences in our tests and examples, but it created
dangerous patterns (as was the case in wrap_fail before the fix) where
tactics raised were caught as part of the CPS computation. In particular
the final raising of a soft failure by eval_judgment was unintentionally
caught by re_raise_tac. This commit also rewrites eval_judgment to limit
the risk for such situations in the future.