Fix handling of errors in checkfail
This follows the previous two commits: checkfail did not capture soft
failures at the right level, which violates CPS and only worked because
of other violations. This is now fixed (together with an erroneous test)
but I left a CPS violation for hard failures, which seems unlikely to
have the same consequences as the previous issues. Still it would be
better to check for hard failures at toplevel only and in direct style
rather than in the middle of the CPS computation of tactics.