fix: use lambda rather than sequences in `fa`
Sequences have no defined semantics on types which are not enumerable
in polynomial time, as opposed to lambdas.

It remains the issues of making sure that these lambdas are on terms
of order at-most 1, which is a known issue that remains to be
fixed (and was already present when using sequences, so this commit
does not introduce any new problem).