refactor: remove outdated asserts and tests
The tests in `tests/ocaml/tactics.ml` verified that the `namelength`
axioms appied to any system: these tests relied on the
`SystemExpr.Any` system, which will be removed soon.
I did not replace these tests, as they only check that the axioms are
general enough (i.e. they are about completeness and not soundness),
and because we already use (and thus test) these axioms extensively
elsewhere in the code.

Also, the asserts in `hyps.ml` will be complex to adapt to the
next way of handling `any`: I thus choose to remove them.