fix: remove redundant tests in `Occurrences` user
This is already tested in the generic occurrences iterators
`Occurrencs.find_all_occurrences` and
`Occurrences.get_macro_actions`.

The precise sub-terms to be ignored depends on the `mode` argument,
which can be either `Const`, `PTimeSI` or `PTimeNoSI` depending on the
tactic being used (e.g. `fresh` requires `const`, and `euf` requires
`PTimeNoSI`).