feat(crypto): better set membership test
(using unification for condition).

Indeed, when doing a set inclusion test

  `{t | ∀ i: phi} ∈ {t0 | ∀ j: phi0}`,

unification was only used to check inference a value of `j` (given by
a substitution `σ`) such that `t = t0σ`, but not to check that we have
`phi → phi0 σ`.

See the second lemma of `tests/ok/crypto-var-inference.sp` for an
example of a `crypto` application that used to fail (because `i = j`
could not be inferred).