fix(crypto): add more general `extra_outputs`
When computing `t` using an oracle, we split the computation of `t`
between a computation of `t | phi` obtained through an oracle call,
and the computation of `t | not phi` obtained directly.

We use to only add `t | phi` to the extra outputs, which is a bad
heuristic: we now `t`, as we know we computed it and it will be easier
to use later one.