fix(crypto): remove assert that was not precise enough
This assert use to fail because it was not precise enough, as
`rec_predicate` may be a conjunction `phi /\ psi` which appears
separated in the list `term.conds`,
e.g. `term.conds = phi :: psi :: ...`.
Since this assert was only here to check that `crypto` is precise and
is not related to soundness, I decided to remove it.