Fix soundness issues with smt and Macros.unfold
SMT translation was incorrect for mutable cells. The translation is now
based on a generic treatment of macros, which fixes the issue.
Hard-coded macro-specific translation cases and SMT theory have been
removed. The change also fixes issues with projections.

When used on mutable cell macro with an arbitrary timestamp,
Macros.unfold did not include cases for actions that do not update the
action.

Update smt steps limits in examples and fix one example to follow
actual macro definitions.

Co-authored with David Baelde.