feat(quantum): quantum deduction (used in `deduce`, `apply`, ...)
Fix the `assumption` tactic in quantum mode, as it was
implicitely applying the classical duplication rule (which is not
sound in quantum mode).

Simplified `fa_decompose`: unification state are now built
afterward and outside the function rather than inside.

Deduction now checks that the order of the terms involved, in
classical and quantum mode. This may break some existing
developements, notably:

- If they used user-declared types without a serializability
  assumption. If applicable, the assumption can be added using the
  `serializable` type restriction. E.g. `type key[large].` may need to
  become `type key[large, serializable]`.

- If they used polymorphic types, e.g. to prove polymorphic deduction
  lemmas. Such lemmas cannot be easily fixed because we currently have
  no way of restricting type variables instantiations. As a temporary
  solution, it is possible to disable order guards in the deduction
  engine using `set deductionOrderGuard=false` (note that doing so could
  lead to unsoundness if incorrectly used).