feat(quantum): temporary flag to disable PQ-checks while transitioning
While we are transitioning tactics one-by-one to be PQ-sound, add
a (internal) flag allowing to by-pass the checks on `qatt`.

The flag default to a PQ-sound value, and is manually set where
needed.

The flag will be removed altogether when we are done.

Currently, used in the `crypto` tactic as well as legacy cryptographic
tactics (i.e. `prf`, `cca`, ...).