Improve timeout mechanism, use it for expandall
Add timeout for expandall tactic. That tactic repeatedly calls
constraint solving and, although each of these calls is subject to
a timeout, the high number of call could cause the whole expandall
process to run way past the timeout. This is now fixed with a timeout
added at the level of expandall in addition to the ones on constraint
solving.

The old timeout mechanism was based on sigalrm, and could not be used
for nested timeouts as required here: calling expandall would setup an
alarm handler and start timer using Unix.alarm; a nested call to
constraint solving would setup another (identical) handler and start
a new timer; once that nested call returns the old handler is
re-installed but the timer is set to 0; effectively, the old timer has
been cancelled.

This commit installs a new timeout mechanism that relies on GC alarms.
It won't be effective for code that does not allocate much, but works
fine in the current use cases. Timeouts are still based on wallclock
time, but this change brings closer the possibility to set the timeout
in terms of allocations, which would be less sensitive to hardward
performances.