fix(crypto): check that memoization hints are inductive
We use to (incorrectly) check this only in time-sensitive mode, even
though memoization is enabled by default in both the standard and the
time-sensitive mode.

There is a single test file were memoization hints are not invariants,
because the memoization heuristic fails to produce an inductive
bi-deduction goal. We added an option `no_memoization` to turn-off
memoization in those cases.