fix: ptime verifications in legacy crypto applications
By adding checks that terms that need to be simulated in ptime do not
contain quantum values, we fix an older unsoundness issue, where it
was not checked that such terms did not contain function symbols of a
type depending on non instantiated type variables that might not be
serializable.

If proofs are broken by this new check, it can probably be fixed by
adding the [serializable] tag to the needed types. (see the update of
some tests in this commit)